feat: introduce CampaignRepository
