+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/campaign/campaign_category.dart
