+++ catalyst_voices/packages/internal/catalyst_voices_repositories/lib/src/catalyst_voices_repositories.dart
@@ -1,4 +1,5 @@
 export 'authentication_repository.dart';
+export 'campaign/campaign_repository.dart';
 export 'credentials_storage_repository.dart';
 export 'proposal/proposal_repository.dart';
 export 'transaction/transaction_config_repository.dart';
