+++ catalyst_voices/packages/internal/catalyst_voices_view_models/lib/src/catalyst_voices_view_models.dart
@@ -1,8 +1,6 @@
 export 'authentication/authentication.dart';
-export 'campaign/campaign.dart';
-export 'campaign/campaign_category.dart';
+export 'campaign/campaign_category_section.dart';
 export 'campaign/campaign_list_item.dart';
-export 'campaign/campaign_section.dart';
 export 'exception/localized_exception.dart';
 export 'exception/localized_unknown_exception.dart';
 export 'menu/menu_item.dart';
